Add JsonSchemaValidator
